✨ feat: ColumnSize enums reference complete

+++ database/migrations/2025_04_01_164811_create_sales_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('users_id');
             $table->unsignedBigInteger('inventories_id');
-            $table->double('price')->unsigned();
+            $table->decimal('price')->unsigned();
             $table->unsignedInteger('quantity');
             $table->timestamps();
         });
